[Bronze V] Title: 이상한 기호, Time: 116 ms, Memory: 20996 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.kt
@@ -1,6 +1,6 @@
 import kotlin.math.*
 
 fun main() {
-    val (a, b) = readln().split(" ").map { it.toInt() }
+    val (a, b) = readln().split(" ").map { it.toLong() }
     println((a+b)*(a-b))
 }
